ENH: set master and merge volumes at once
This avoids the check for the node name in set merge node, which is currently
not working as expected, and as a result gets rid of the Editor popup window
while setting the merge node (see proposed soultion https://github.com/Slicer/Slicer/pull/263).

Thanks to @rweiss42 for investigating the issue, and to @pieper for the
setVolumes() hint!